Moved database entities into same file
